// src/model/accept-apply-info.d.ts
export interface AcceptApplyInfo {
  codeAcceptApply: string;//验收单编码
  stateAcceptApply: string;//验收单状态
  contractCodeAcceptApply: string;//合同流水号
  contractNameAcceptApply: string;//合同名称
  projTypeAcceptApply: string;//项目性质
  projPartNameAcceptApply: string;//项目单元名称
  applyDepartAcceptApply: string;//申请单位
  applyTimeAcceptApply: string;//申请时间
  applyAcceptApply: string;//申请人
}